Add pe-mcollective-hotfix-service and refactor
Without this patch the `pe_280_mcollective_fix` module will not compile
in puppet apply mode because of a missing `pe-mcollective` service
resource. This is only an issue in puppet apply mode, things works fine
if the `pe-mcollective` class is included earlier in the catalog.

The pe_280_mcollective_fix module has also been refactored to follow the
params.pp pattern and adhere to the Puppet Labs style guide.

A Modulefile has been added which makes it easy to build a Puppet module
package and host it on the Puppet Forge. To build the Puppet module
package, change into the source directory and run the following command:

    $ puppet module build .
    Building /root/pe_280_mcollective_fix for release
    pkg/puppetlabs-pe_280_mcollective_fix-0.0.1.tar.gz

Once the package has been built and uploaded to the Forge, run the
following command on the master or agent -- if using puppet apply mode,
to install the module:

    $ puppet module install pkg/puppetlabs-pe_280_mcollective_fix

With the module installed a node can be classified from the PE console
with the pe_280_mcollective_fix class. The next Puppet run should apply
the pe_280_mcollective_fix and Live Management should be fix!

The module also works in puppet apply mode:

    $ puppet apply -e "include pe_280_mcollective_fix"

Tested on Debain Squeeze running PE 2.8.0.