fixes duplicate package issues for rhel
also a first pass on fixing this for debian systems.

params.pp adds stomp_pkg which is the full package namen, and then uses
stomp_name to refer to the package that should be installed.

folks more familiar with solaris and debian package providers should be
able to do the same to make sure it works for those.