A fix for Solaris, we think
This hasn't been tested yet, and is more fragile than other fixes.

We need to install the package with -G, and unfortunately the ability to do so
wasn't added to the sun provider until Puppet 3.0. So we need to do it as an
exec instead of a package. This commit adds an if statement so we handle this on
Solaris, adds an exec, adds an admin file for that exec, and changes the apply
wrapper to not depend on the package resource that now won't be present on
Solaris. (It should still work fine, since it's being notified by whatever needs
to notify it.)